Cleaned up make targets. Moved cidr utils to iac-shared. Update docs
