+++ iac/iac-shared/package.json
@@ -48,7 +48,8 @@
     "@pulumi/command": "^0.8.2",
     "@pulumi/postgresql": "^3.9.0",
     "@pulumi/pulumi": "^3.79.0",
-    "@types/node": "^20.5.6"
+    "@types/node": "^20.5.6",
+    "cdktf": "^0.18.0"
   },
   "devDependencies": {
     "@ava/typescript": "^4.1.0",
