+++ iac/iac-shared/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./lib/hashing";
+export * from "./lib/utils";
 
 export {AccountEntity, AccountConfig, VpcConfig} from 'AccountSchema';
 export * from './lib/AccountSchema';
