+++ makefile
@@ -43,23 +43,3 @@ build:
 watch-lambda:
 	bin/watchman.sh $(PROJECT_MODULE_NAME) "make build"
 
-test-cdktf:
-	make local-cdktf-output ARGS="--outputs-file ../../../auto_tests/cdktf-output.json"
-	cd auto_tests && jq '."LsLambdaS3Sample.local"' cdktf-output.json > iac-output.json;
-	make test-cdktf-bare
-
-test-cdktf-bare:
-	$(VENV_RUN) && cd auto_tests && AWS_PROFILE=localstack pytest $(ARGS);
-
-
-reset:
-	- stop-ls.sh
-	- rm iac/Pulumi.*.yaml
-	- rm -rf ./global-iac/.pulumi
-	- rm -rf iac/.pulumi
-	- start-ls.sh
-
-it-again: reset
-	make local-toplevel-deploy
-	make local-ecr-deploy
-
